+++ src/include/LSP/Utils.hpp
@@ -15,6 +15,8 @@ void trim_end(std::string& str);
 void trim(std::string& str);
 std::string& toLower(std::string& str);
 bool endsWith(const std::string_view& str, const std::string_view& suffix);
+bool replace(std::string& str, const std::string& from, const std::string& to);
+void replaceAll(std::string& str, const std::string& from, const std::string& to);
 
 template<class K, class V>
 inline bool contains(const std::unordered_map<K, V>& map, const K& value)