+++ CHANGELOG.md
@@ -9,6 +9,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Added `luau-lsp.sourcemap.enabled` option which dictates whether sourcemap-related features are enabled
+- Diagnostics will now be provided for definitions files which errored
+
+### Changed
+
+- Moved definitions file loading to post-initialization
 
 ## [1.5.2] - 2022-06-22